Outdated Project Plan.
Clearly, I did Indeed code the actual thing after the skeleton was made, mostly because there was only so many days left to build it. I was becoming overly obsessed with making the perfect system that I risked making no actual progress, what with all of the restarting. Ultimately, I chose to focus on the game instead of the engine.